rework KATIE_SETUP_SOURCES CMake macro
now that the assembler language is enabled in the build system the macro
turned out to be a bit bogus and it has to handle sources control too
because some CPU specifiec sources were, previously, added unconditionally
to targets leading to build failures (especially with assembly).

CPU tests now declare variables that are cached and can be changed by the
user to avoid touching parent scope from the test macros (previously
functions for that reason).

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>