add shells profile file to workaround musl dynamic loader issue
this file guarantees that on hardened musl libc systems the
libraries can be found by the libraries loader when build without
RPATH and the custom ld config is not respected. it does require
a logout to take effect but I do not know what else can be done
about that.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>