no need for SIMD in imageconversions and drawing now
now that image format is not changed when scaling the performance
impact with this change is almost none, in most cases you will
(should) not notice a difference. if you care about performance
you most likely build with -mtune=native anyway so that the
compiler can better optimize things for the hardware.

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>