remove the gray raster code
this is mostly for test purposes only. the removal will affect
some of the widgets drawing, especially round corners (like the
Plasma "Tool Box" and radio buttons with the Oxygen style). with
previous tests some changes to the outline mapper (QOutlineMapper)
were needed to fix some rounding issues but caused others so I am
not sure if I can make it work well without the Freetype2
rasterizer but I sure will try, again, if I have the time for it.

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>