reimplement QRegion via X11 provided functions
it works for regions without extents, i.e. regions that are not in the
bounds are not redrawn properly which probably means Xfixes extension
has to be used to mark damaged areas however that has yet to be
tested and it may require slight changes in the the way rendering is
done.

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>