do not append application path to library paths
this was used for dll/bundles support most likely, on UNIX however adding
it to the library paths is incorrect. The runtime path (RPATH) hardcoded
into libraries/binaries should be searched instead however that is not
supported and removed by most distributions on purpose.

with this change application startup time is reduced by a lot since the
binary paths (/bin, /sbin, etc.) do not have to be listed which usually
contain large number of binaries and even symbolic/hard links

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>