fix loading of plugins with a relative file name
the libraries path was missing from the list returned by
QCoreApplication::libraryPaths() which as stated in its documentation
is used when dynamically loading libraries so that's also fixed with this
change and is required for the libraries lookup.

obviously, QLibraryPrivate::findOrCreate() was not looking for libraries at
all despite the function name so code for that had to be moved from
QLibraryPrivate::load_sys() which makes it possible to load not only plugins
but also library with relative file name.

support for QLibrary::LoadHint::LoadArchiveMemberHint was dropped in the
process, QLibrary::LoadHint::ImprovedSearchHeuristics is not used thus
removed.

in the future QCoreApplication::pluginPaths() may be added to separate the
library/plugin paths but for now compatibility with projects expecting it
to return plugin paths will be kept. when that is done
QCoreApplication::libraryPaths() will check LD_LIBRARY_PATH instead of
QT_PLUGIN_PATH environment variable.

upstream commit:
https://github.com/qt/qtbase/commit/418890e0748384eb684f33b10dc6f32493aee54b

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>