rework the translation related functions
this is both step back (in terms of plural support) and step forward
(in terms of missuse of QApplication::translate()) but it has to be done
to make the API somewhat sensible. note that plural forms emulation via
QString.arg() is already done in some places and
QApplication::translate() is used somewhere where it should not be. in a
nutshell, if your code does not compile it probably needs more than a
build fix

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>