rendering will not be done based on script
as it is done now, fonts are probed for character specific to script (see
qfontdatabase source file) which assumes that if a font can render one
character for a given script it can render all characters for that script -
this is very much not so. fonts usually are incomplete as new scripts come
and go fonts have to be updated constantly to account for that and some
of the work done on fonts is incomplete. instead fonts have to be probed
for the character that is to be rendered

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>